# server/openapi.yaml
openapi: 3.0.3
info:
  title: ActivityOS API
  version: 0.31.0
  description: |
    Express + Firestore backend for ActivityOS — a **multi-tenant** platform
    for children's activity providers.

    ## v0.31.0 — Cross-provider marketplace (opt-in)

    The parent Browse feed becomes a real marketplace instead of being scoped
    (client-side) to providers a family had already booked.
    - **`GET /api/listings`** (no `?tenantId=`, no `?mine=1`) now returns the
      live/public/unarchived listings of every provider **that has opted in**
      (`settings.marketplaceListed === true`) — **plus**, for a signed-in
      parent, their own providers (anyone they've booked) regardless of opt-in.
    - A provider's **own storefront** (`GET /api/listings?tenantId=`) is
      unchanged — their public listings always show there whether or not
      they're in the marketplace.
    - New tenant setting **`marketplaceListed`** (in the library `settings`
      bag), default **false** — a provider opts in from Setup → Bookings →
      Marketplace. Opt-in is a deliberate choice; nobody is publicly listed
      without it.
    (Frontend: BrowseApp drops the old single-provider scoping and shows the
    server feed as-is.)

    ## v0.30.0 — Account, Data & privacy, parent Accidents

    Clearing the shared/parent mock pages.
    - **`/api/account`** (every role) — the signed-in user's profile. `GET`
      returns `{email, name, phone, marketingConsent, role, tenantId}` (identity
...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - firebaseIdToken: []
paths:
...
  /api/shifts:
    get:
      tags: [shifts]
      summary: The rota (staff read too); window with from/to
      parameters:
        - { name: from, in: query, required: false, schema: { type: string }, description: "ISO date, inclusive" }
        - { name: to, in: query, required: false, schema: { type: string }, description: "ISO date, inclusive" }
      responses:
        "200": { description: Array of shifts, sorted by date then start }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [shifts]
      summary: Add a shift (operators only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [staffName, date, start, end]
              properties:
                staffName: { type: string }
                staffId: { type: string }
                date: { type: string, description: "ISO date" }
                start: { type: string, description: "HH:MM" }
                end: { type: string, description: "HH:MM" }
                role: { type: string }
                listingId: { type: string }
                notes: { type: string }
      responses:
        "201": { description: The created shift }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { description: "Blocked by Settings → Staff & workforce: requireDBS (missing/expired DBS) or requireCompliance (expired first aid / safeguarding cert). Only enforced once the certifications register has entries." }
  /api/shifts/{id}:
    put:
      tags: [shifts]
      summary: Edit a shift (operators only)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "200": { description: The updated shift }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [shifts]
      summary: Delete a shift (operators only)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: "{ok:true}" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/ratios:
    get:
      tags: [ratios]
      summary: The day's sessions with staffing ratios and groups
      parameters:
        - name: date
          in: query
          required: false
          schema: { type: string, example: "2026-08-03" }
        - name: tenantId
          in: query
          required: false
          description: Platform accounts only
          schema: { type: string }
      responses:
        "200":
          description: "{date, bands, sessions[]} — each session: children, requiredStaff, staffAssigned, met, groups"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/ratios/board/{date}:
    get:
      tags: [ratios]
      summary: The day's cover board (drag overrides + per-group staffing)
      description: One board per tenant per day — the state the Ratios cover board renders. Staff and operators.
      parameters:
        - name: date
          in: path
          required: true
          schema: { type: string, example: "2026-08-03" }
      responses:
        "200": { description: "{overrides: {childKey: groupId}, groupStaff: {groupId: staffIds[]}}" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    put:
      tags: [ratios]
      summary: Save the day's cover board (staff + operators)
      parameters:
        - name: date
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                overrides:
                  type: object
                  description: childKey → groupId (a manual drag out of the age-derived group)
                groupStaff:
                  type: object
                  description: groupId → array of staff ids covering it
      responses:
        "200": { description: Saved }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/ratios/{blockId}/{date}:
    put:
      tags: [ratios]
      summary: Save a session's groups + staffing (operators)
      parameters:
        - name: blockId
          in: path
          required: true
          schema: { type: string }
        - name: date
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [groups]
              properties:
                groups:
                  type: array
                  items:
                    type: object
                    properties:
                      id: { type: string }
                      name: { type: string }
                      childIds: { type: array, items: { type: string } }
                      staffIds: { type: array, items: { type: string } }
      responses:
        "200": { description: Saved }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    firebaseIdToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token (see the Authentication section above).
...
  responses:
    BadRequest:
      description: Validation failed (`error` is a message or zod issue list)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing/invalid Firebase ID token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: The account's role/tenant doesn't allow this
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found (or outside the caller's tenant scope)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      properties:
        error:
          description: Human-readable message, or an array of zod issues on 400.
          oneOf:
            - type: string
            - type: array
              items: { type: object }
...